+++ cinelerra-4.6.mod/cinelerra/pipe.C
@@ -96,7 +96,7 @@ int Pipe::substitute()
 		}
 
 		// insert the file string at the substitution point
-		if (f + strlen(sub_str) - complete > sizeof(complete))
+		if (f + strlen(sub_str) - complete > (int)sizeof(complete))
 		{
 			printf("Pipe::substitute(): max length exceeded\n");
 			return -1;
